MLOPS_FINAL_COURSE
The main part of this proyect is explain the fluctuations in the duration of the trips of the yellow taxis of the NY city during the first 3 months of 2022.

New York is one of the biggest cities around the world, therefore, known about the duration of the trips is an excellent way to analyze the vehicle movements in the city and what are the variables that explains better that results.

For this model, was developed a Linear Regression model that has 5 categorical variables that are the locations, price, passengers_count and trip distance.